backend-windows/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
mss
pillow
pyperclip
psutil
pydantic
python-multipart
requests
bcrypt==4.0.1
pywin32
